isolate constants for build size
