+++ src/common/constants/agent-constants.js
@@ -15,19 +15,3 @@ export const SUPPORTS_REGISTERED_ENTITIES = {
   [FEATURE_NAMES.jserrors]: true,
   [FEATURE_NAMES.ajax]: true
 }
-
-export const EVENT_TYPES = {
-  AJAX: 'AjaxRequest',
-  PA: 'PageAction',
-  UA: 'UserAction',
-  BP: 'BrowserPerformance',
-  WS: 'WebSocket',
-  SPV: 'SecurityPolicyViolation',
-  JSE: 'JavaScriptError',
-  LOG: 'Log',
-  PVE: 'PageView',
-  PVT: 'PageViewTiming',
-  SR: 'SessionReplay',
-  ST: 'SessionTrace',
-  BI: 'BrowserInteraction'
-}
